Split the classes from EventHubsTestUtils.scala in seperate files
The EventHubsTestUtils.scala was very long, therefore it is better
to split the classes/objects into seperate files
